保存一下

+++ run.py
@@ -214,13 +214,21 @@ def test_wrapper(model):
     trainer.test(model, test_input_handle, args, 'test_result')
 
 
-if os.path.exists(args.save_dir):
-    shutil.rmtree(args.save_dir)
-os.makedirs(args.save_dir)
+if args.is_training:
+    if os.path.exists(args.save_dir):
+        shutil.rmtree(args.save_dir)
+    os.makedirs(args.save_dir)
 
-if os.path.exists(args.gen_frm_dir):
-    shutil.rmtree(args.gen_frm_dir)
-os.makedirs(args.gen_frm_dir)
+    if os.path.exists(args.gen_frm_dir):
+        shutil.rmtree(args.gen_frm_dir)
+    os.makedirs(args.gen_frm_dir)
+else:
+    if not os.path.exists(args.save_dir):
+        os.makedirs(args.save_dir)
+
+    if not os.path.exists(args.gen_frm_dir):
+        os.makedirs(args.gen_frm_dir)
+
 
 print('Initializing models')
